requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.7.1+cu110
transformers==3.5.1
protobuf==3.20.*
openpyxl==3.0.9
PyYAML==6.0
selenium==3.141.0
selenium-wire==1.0.8
webdriver_manager==3.4.2